+++ src/sources/source_typedtable.jl
@@ -12,7 +12,8 @@ function query(df::TypedTables.Table)
     col_expressions = Array{Expr,1}()
     df_columns_tuple_type = Expr(:curly, :Tuple)
     for i in 1:length(df.data)
-        push!(col_expressions, Expr(:(::), names(df)[i], eltype(df.data[i])))
+        etype = eltype(df.data[i])
+        push!(col_expressions, Expr(:(::), names(df)[i], etype <: Nullable ? NAable{etype.parameters[1]} : etype))
         push!(df_columns_tuple_type.args, typeof(df.data[i]))
     end
     t_expr = NamedTuples.make_tuple(col_expressions)
@@ -43,8 +44,8 @@ end
 
 @generated function next{T,TS}(iter::EnumerableTypedTable{T,TS}, state)
     constructor_call = Expr(:call, :($T))
-    for i in 1:length(iter.types[2].types)
-        push!(constructor_call.args, :(columns[$i][i]))
+    for (i,t) in enumerate(T.parameters)
+        push!(constructor_call.args, t<:NAable ? :(isnull(columns[$i][i]) ? NAable{$(t.parameters[1])}() : NAable{$(t.parameters[1])}(get(columns[$i][i]))) : :(columns[$i][i]))
     end
 
     quote