+++ NEWS.md
@@ -1,3 +1,7 @@
+Query.jl v.next Release Notes
+=============================
+* Use NAable instead of Nullable
+
 Query.jl v0.3.0 Release Notes
 =============================
 * Add Dict sink support